% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dgt_image.R
\name{read_dgt_image}
\alias{read_dgt_image}
\title{Read a two-dimensional DGT intensity image}
\usage{
read_dgt_image(path, invert = FALSE, allow_color = FALSE, rgb_tolerance = 0.02)
}
\arguments{
\item{path}{Path to a PNG or JPEG image.}

\item{invert}{Whether to replace intensity \code{x} with \code{1 - x}.}

\item{allow_color}{Whether to permit luminance conversion of a colored RGB image.}

\item{rgb_tolerance}{Maximum channel difference treated as grayscale.}
}
\value{
A numeric intensity matrix with values in \verb{[0, 1]}.
}
\description{
Read a two-dimensional DGT intensity image
}
